OCI: fix signed vs unsigned comparison (refs https://trac.osgeo.org/gdal/ticket/6866)
git-svn-id: https://svn.osgeo.org/gdal/trunk@38058 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/oci/ogr_oci.h
@@ -129,7 +129,7 @@ class CPL_DLL OGROCISession {
 
     int         nServerVersion;
     int         nServerRelease;
-    int         nMaxNameLength;
+    size_t      nMaxNameLength;
 };
 
 OGROCISession CPL_DLL*
